requirements.txt
Django==1.4
PIL==1.1.7
Pygments==1.6
South==0.7.6
argparse==1.2.1
beautifulsoup4==4.1.3
boto==2.8.0
dj-database-url==0.2.1
django-debug-toolbar==0.9.4
django-extensions==1.0.3
-e git+git://github.com/toastdriven/django-haystack.git#egg=haystack
django-registration==0.8
django-social-auth==0.7.22
django-storages==1.1.6
docutils==0.10
eventbrite==0.45
html5lib==0.95
httplib2==0.8
ipython==0.13.1
oauth2==1.5.211
psycopg2==2.4.6
python-dateutil==2.1
python-openid==2.2.5
python-ptrace==0.6.4
pytz==2012j
simplejson==2.6.1
six==1.3.0
sorl-thumbnail==11.12
twill==0.9
wsgiref==0.1.2
pygithub
django-endless-pagination==2.0
